A few updates to support the tool version
Added offsets to touch position
Added sending WM_NOTIFY
Added some OpenGL utility code

+++ source/display/prOglUtils.cpp
@@ -89,7 +89,7 @@ void prOpenGLInit()
     {
         initialised = true;
 
-        PRGL_VERSION = atof((const char *)glGetString(GL_VERSION));
+        PRGL_VERSION = (f32)atof((const char *)glGetString(GL_VERSION));
         prTrace("OpenGL version %f\n", PRGL_VERSION);
     }
 }
