+++ source/display/prTrueTypeFont.cpp
@@ -215,12 +215,12 @@ void make_dlist(FT_Face face, char ch, GLuint list_base, GLuint *tex_base )
 	//first we need to move over a little so that
 	//the character has the right amount of space
 	//between it and the one before it.
-	glTranslatef(bitmap_glyph->left, 0, 0);
+	glTranslatef((GLfloat)bitmap_glyph->left, 0, 0);
 
 	//Now we move down a little in the case that the
 	//bitmap extends past the bottom of the line 
 	//(this is only true for characters like 'g' or 'y'.
-	glTranslatef(0, bitmap_glyph->top-bitmap.rows, 0);
+	glTranslatef(0, (GLfloat)bitmap_glyph->top-bitmap.rows, 0);
 
 	//Now we need to account for the fact that many of
 	//our textures are filled with empty padding space.
@@ -242,13 +242,13 @@ void make_dlist(FT_Face face, char ch, GLuint list_base, GLuint *tex_base )
 	//glTexCoord2d(0,y); glVertex2f(0,0);
 	//glTexCoord2d(x,y); glVertex2f(bitmap.width,0);
 	//glTexCoord2d(x,0); glVertex2f(bitmap.width,bitmap.rows);
-	glTexCoord2d(x,y); glVertex2f(bitmap.width, bitmap.rows);
-	glTexCoord2d(x,0); glVertex2f(bitmap.width, 0);
+	glTexCoord2d(x,y); glVertex2f((GLfloat)bitmap.width, (GLfloat)bitmap.rows);
+	glTexCoord2d(x,0); glVertex2f((GLfloat)bitmap.width, 0);
 	glTexCoord2d(0,0); glVertex2f(0, 0);
-	glTexCoord2d(0,y); glVertex2f(0, bitmap.rows);
+	glTexCoord2d(0,y); glVertex2f(0, (GLfloat)bitmap.rows);
 	glEnd();
 	glPopMatrix();
-	glTranslatef(face->glyph->advance.x >> 6 ,0,0);
+	glTranslatef((GLfloat)(face->glyph->advance.x >> 6), 0, 0);
 
 
 	//increment the raster position as if we were a bitmap font.
